backend/requirements.txt
Django==4.0.6
django-cors-headers==3.13.0
djangorestframework==3.13.1
drf-yasg==1.21.0
PyJWT == 2.5.0
